made new file for the chatScreenPeoplePane
