+++ packages/react-composites/src/composites/ChatComposite/ChatScreen.tsx
@@ -31,14 +31,8 @@ import {
   typingIndicatorChatCompositeStyles,
   typingIndicatorContainerStyles
 } from './styles/Chat.styles';
-
-/* @conditional-compile-remove-from(stable) */
-import { ParticipantContainer } from '../common/ParticipantContainer';
-/* @conditional-compile-remove-from(stable) */
-import { useLocale } from '../localization';
 import { participantListContainerPadding } from '../common/styles/ParticipantContainer.styles';
-/* @conditional-compile-remove-from(stable) */
-import { ParticipantList } from '@internal/react-components';
+import { ChatScreenPeoplePane } from './ChatScreenPeoplePane';
 
 /**
  * @private
@@ -73,11 +67,6 @@ export const ChatScreen = (props: ChatScreenProps): JSX.Element => {
 
   const adapter = useAdapter();
 
-  /* @conditional-compile-remove-from(stable) */
-  const locale = useLocale();
-  /* @conditional-compile-remove-from(stable) */
-  const chatListHeader = locale.strings.chat.chatListHeader;
-
   useEffect(() => {
     adapter.fetchInitialData();
   }, [adapter]);
@@ -87,8 +76,6 @@ export const ChatScreen = (props: ChatScreenProps): JSX.Element => {
   const typingIndicatorProps = usePropsFor(TypingIndicator);
   const headerProps = useAdaptedSelector(getHeaderProps);
   const errorBarProps = usePropsFor(ErrorBar);
-  /* @conditional-compile-remove-from(stable) */
-  const participantListProps = usePropsFor(ParticipantList);
 
   const onRenderAvatarCallback = useCallback(
     (userId, defaultOptions) => {
@@ -135,11 +122,9 @@ export const ChatScreen = (props: ChatScreenProps): JSX.Element => {
         {
           /* @conditional-compile-remove-from(stable) */
           options?.participantPane === true && (
-            <ParticipantContainer
-              participantListProps={participantListProps}
+            <ChatScreenPeoplePane
               onFetchAvatarPersonaData={onFetchAvatarPersonaData}
               onFetchParticipantMenuItems={props.onFetchParticipantMenuItems}
-              title={chatListHeader}
             />
           )
         }
